refactor: use seperate files for service and types
